+++ doc/version_history.rst
@@ -13,6 +13,7 @@ v7.8.0
 ------
 
 * Added optional ``stem`` parameter to ``AsyncS3Bucket::make_key``.
+* Implemeted URL return from ``AsyncS3Bucket::upload``.
 
 
 v7.7.1